fixed jankyness with suggestions/autocomplete on slow computers
if a user on a slower pc were to use the suggestion function they would
experience the cursor moving while the characters of the suggestions are
being typed. this was ugly and looked slow.

fixed this by adding a buffer of the sugestions and being sure that the
whole suggestion is typed out before being shown to the user. This makes
the suggestion a little delayed on slow pc's but yields a better
experience. the speed should be the same on higher end pc's becuase it
displays as fast as its typed the suggested characters. there is no set
delay.

this makes the feature only as fast as the computer, which makes more
sense