added primative history command
typing "history" into the shell will display a list of the last 10
commands the user entered into the shell, and enumurate them.

this is a very primative implementation and cycling through this list is
not possible yet. it is only good for displaying them

- Features getting added -

going to add a function to hit tab to cycle through the list and hit
ENTER to accept them and run the current command