small fixes and new features
fixed some small problems with autocomplete

when the user started typing, it would only suggest files and folder,
not commands. this does not make sense, as in 99% of cases the user will
type the command first. it will now suggest a command as the first word

for the 1% of the time for cmd users to execute a command with just the
file name, the suggestion will change to file suggestion if the first
couple of letters do not match any commands