new features and fixes
-- suggestions --

minor changes to the suggestion command. before this update if the user
had 2 similar folders and wanted to cycle through them they would have
to hit tab to enter cycling mode and then tab again to actually cycle to
the next entry. this is now updated to when the user presses tab it
enters cycle mode and also cycles once

Example: say i have these two folders "shell" "shellTest" and i type "cd
sh" now the suggestion would suggest shell, if i now hit enter i would
see shellTest and be in cycling mode

also removed case sensitivity in suggestion. so if i have this folder
"FolderExample" the suggestion could not find the folder if i wrote "cd
fol..." now it works and is case insensitive

-- ls/dir --

added feature where the header row is added to the bottom if the list of
files and folders is longer than the window. this will ensure that the
user will always be aware of what they are looking at without having to
scroll all the way up

-- thanks to NuShell for this idea. great work!