feature - Status line
added the status line to the shell. this status line is permanently at
the bottom of the screen, and will not move, this ensures that all your
most important details are always accessable at a glance

moved the git info to the status line

before the git info was attached to the prompt itself, and the user
would often end up with a wall of text on the right of the exact same
info. this is now fixed by moving the git info to the status line.
ensuring that it is only changed if the user changes a directory