updt: changed the suggestion handling
to accept a suggestion use tab. using enter was causing a lot of
problems, especially where the users would get stuck in "suggestions
hell". this is now fixed by making tab the accept button.

this also change the behabiour of cycling through suggestion. it now
completes the first suggestion on the first hit of TAB and second hit of
tab will enter cycle mode