feature! added a timer
added a timer feature that people can use for concentration

Usage: timer h m s.

will peep when done and all you have to do is cls in the shell. if you
are coding it will not hinder any key strokes so it is easier to use.

this solution is half baked and needs updates i know. some known issues
are the cursor is redrawing when the timer is redrawing which is ugly
and annoying. the timer is blocking for things like lualine and similar
when in neovim. the stop message also breaks if the neovim window
scrolls

this is all getting fixed in the future, it is low on prioreties right
now but if asked for i can work on it

shoot me a dm on github!