several updates to the timer functin
the timer command would not accept minutes and hours as input. now this
works

the custom session name now also works. Usage: timer 1h20m "vibe coding"

the notification box has been semi-fixed. theres still some problems
with closing it and im working on it. though it is not high on my list
of priorities.

if you for some weird reason really need this to work ASAP, shoot me a
msg on github and ill fix it. (or clone the repo and fix it youself)