updates to clearing handling
previously the cls/clear command would only clear whats in view, which
would make it seem like it was working, but if the user scrolls up using
the mouse they can see the parts of prevoius commands that wasnt in the
view port during the clearing function.

updated to just use system("cls") which worked on the 3 terminals i
tried it on on windows (warp, wezterm and windows terminal) i did leave
in a backup function which can clear nativly but it is commented out, so
either dm me on github and i can uncomment it or uncomment it yourself
in builtins.c lsh_clear() function.