reworked entire grep view
the grep view now actually works and looks like intented. fixed some
drawing bugs and missing lines in the list of files.

-- Feature --

when in grep and you found the instance of your search, instead of
opening your code editor and going to that line, simply hit enter to
open that file on that line on the keyword

when in the normal grep view and you feel like you want to see more of
the code snippet, open a big view by pressing 'o'