updated fzf searching and drawing
searching was really slow before becuase it would load for every keyword

this is now fixed with by having to hit enter for searching. i know this
isnt optimal because you dont remember the file name sometimes

this is getting fixed but it was just so slow wihout it and it just made
it painfull to use.

going to implement a real lazy loading handler to make this better but i
just cant figure it out right now