updated fzf
updated the native fzf to launch with syntax highlighting, this is just
a nice thing to have

also added alot of fallbacks in the last commit but forgot to mention
them

if bat cant be found as a preview type it will default to cat, which
should work in every case. if ripgrep fails or the user doesnt have
ripgrep it will fall back to my own implementation of ripgrep, which is
slower but definately does the job.

theres also a fallback in fzf too, both if the user doesnt have bat, or
fzf entirely. if the user doesnt have bat it will use cat. if the user
doesnt have fzf it will fall back to my own implementation of fzf, which
is not as feature rich but has an ok ui and does its job

if the shell detects that the user is missing anything the shell will
give a short instruction on how to install whats needed

this makes some of the nicest features of the shell accessable for all
even if that means half baked implementations done by a mediocre at best
dev