fixed the suggestions
the history suggestions was messing with the displaying of the
suggestion of the argument type. now typing something that hasnt been
typed before will suggest the argument type.

also changed it so shift + enter accepts any suggestion not just history
suggestions, this makes more sense to my brain, going to add a config
for this later when i have the energy