differentiating suggestions
the system has 2 suggestions. one is the context match (for commands and
argument types) and the other one is for history based suggestion

since these 2 were using the same color it was impossible to know when
to press what. now the history suggestion is the default grey color and
the command / argument type is blue