added alot of shit
added command validation checking before hitting enter. if the command
is not recognized it will light up red. as soon as the command is
recognized it will stop lighting up.

added external command awareness so that when u install new shit the
suggestions will find them, this is abit agressive right now and im
trying to fix it. this also works with the command syntax so easy work